refactor: added use-project-state hook
