+++ common/client/fragments/project-profile-fragment/src/project-profile.component.tsx
@@ -1,9 +1,6 @@
 import React                  from 'react'
 import { FC }                 from 'react'
 import { FormattedMessage }   from 'react-intl'
-import { useRouter }          from 'next/router'
-import { useEffect }          from 'react'
-import { useState }           from 'react'
 
 import { Project }            from '@shared/data'
 import { Condition }          from '@ui/condition'
@@ -14,33 +11,15 @@ import { Layout }             from '@ui/layout'
 import { Text }               from '@ui/text'
 import { Title }              from '@ui/title'
 import { WrapperWhite }       from '@ui/wrapper'
-import { mockAuthUserId }     from '@shared/data'
-import { useGetUser }         from '@shared/data'
 
 import { ButtonEditProject }  from './button-edit-project'
 import { ProfileAvatar }      from './profile-avatar'
 import { ProfileDescription } from './profile-description'
 import { ProfileTeams }       from './profile-teams'
-import { useGetProject }      from './data'
+import { useProjectState }    from './hooks'
 
 export const ProjectProfile: FC = () => {
-  const router = useRouter()
-  const queryId = typeof router.query.id === 'string' ? router.query.id : ''
-
-  const [project, setProject] = useState<Project>()
-
-  const { projectTeams, project: fetchedProject } = useGetProject({ id: queryId })
-  const { user: ownerUser } = useGetUser({ id: project?.ownerUserId || '', skip: 0, take: 20 })
-
-  const [isMyProject, setIsMyProject] = useState(false)
-
-  useEffect(() => {
-    if (fetchedProject) {
-      if (fetchedProject.ownerUserId === mockAuthUserId) setIsMyProject(true)
-
-      setProject(fetchedProject)
-    }
-  }, [router, fetchedProject])
+  const { project, setProject, isMyProject, projectTeams, ownerUser } = useProjectState()
 
   const handleEditProject = (editedProject: Project) => setProject(editedProject)
